updated to lucene-core-2.3.2 and modified DistanceSortSource to return SortField.DOUBLE for sortValue

+++ locallucene/src/java/com/pjaol/search/geo/utils/DistanceSortSource.java
@@ -80,7 +80,7 @@ public class DistanceSortSource implements SortComparatorSource {
 		}
 
 		public int sortType() {
-			return SortField.FLOAT;
+			return SortField.DOUBLE;
 		}
 
 		public Comparable sortValue(ScoreDoc iDoc) {
